fix: reflects the GoMethods to Golistics rename which was to avoid potential naming collisions in golanglint-ci namespace
+ renames the folder and files under `pkg/golinters`.
+ renames the directives in test file accordingly.
+ downgrades the test module Go version to 1.24 as by the submission requirements eg. https://github.com/golangci/golangci-lint/pull/6303#issuecomment-3726807089.
+ updates the name in linter config in testdata.
+ updates the name `lintersdb` entry.
+ updates `go.mod` and `go.sum`
+ reorders list entries by alphabetically.
