+++ pkg/golinters/golistics/golistics.go
@@ -1,9 +1,9 @@
-package gomethods
+package golistics
 
 import (
 	"github.com/golangci/golangci-lint/v2/pkg/goanalysis"
 
-	"go.ufukty.com/gomethods/pkg/analyzer"
+	"go.ufukty.com/golistics/pkg/analyzer"
 )
 
 func New() *goanalysis.Linter {
