+++ pkg/golinters/golistics/golistics_test.go
@@ -1,4 +1,4 @@
-package gomethods
+package golistics
 
 import (
 	"testing"
@@ -6,6 +6,6 @@ import (
 	"github.com/golangci/golangci-lint/v2/test/testshared/integration"
 )
 
-func TestGomethods(t *testing.T) {
+func TestGolistics(t *testing.T) {
 	integration.RunTestdata(t)
 }
