+++ pkg/golinters/golistics/testdata/go.mod
@@ -1,3 +1,3 @@
 module test
 
-go 1.25.5
+go 1.24.0
