+++ pkg/golinters/golistics/testdata/golistics.go
@@ -62,7 +62,7 @@ func collect(_ map[string]any) []string { return nil }
 func tree(_ ...any) string              { return "" }
 func safeEq(a, b any) bool              { return a == b }
 
-//gomethods:exported
+//golistics:exported
 func (s Text) Strings() []string {
 	return collect(map[string]any{
 		"Color":         s.Color,
@@ -71,12 +71,12 @@ func (s Text) Strings() []string {
 	})
 }
 
-//gomethods:exported // want `missing fields: Height, Width`
+//golistics:exported // want `missing fields: Height, Width`
 func (Dimensions) Strings() []string {
 	return nil
 }
 
-//gomethods:exported
+//golistics:exported
 func (s *Styles) Strings() []string {
 	return collect(map[string]any{
 		"Dimensions":      s.Dimensions,
@@ -91,24 +91,24 @@ func (s *Styles) Strings() []string {
 	})
 }
 
-//gomethods:exported
+//golistics:exported
 func (r Rule) String() string {
 	return tree(r.Selector, r.Styles.Strings())
 }
 
-//gomethods:all // want `missing fields: Bottom, Left, Right, Top`
+//golistics:all // want `missing fields: Bottom, Left, Right, Top`
 func (Borders) IsEqual(Borders) bool {
 	return false
 }
 
-//gomethods:all // want `missing field: Top`
+//golistics:all // want `missing field: Top`
 func (s Margin) IsEqual(y Margin) bool {
 	return safeEq(s.Right, y.Right) &&
 		safeEq(s.Bottom, y.Bottom) &&
 		safeEq(s.Left, y.Left)
 }
 
-//gomethods:exported
+//golistics:exported
 func (s Dimensions) IsEqual(y Dimensions) bool {
 	return safeEq(s.Height, y.Height) &&
 		safeEq(s.Width, y.Width)
