+++ pkg/lint/lintersdb/builder_linter.go
@@ -51,7 +51,7 @@ import (
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/goheader"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/goimports"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/golines"
-	"github.com/golangci/golangci-lint/v2/pkg/golinters/gomethods"
+	"github.com/golangci/golangci-lint/v2/pkg/golinters/golistics"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/gomoddirectives"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/gomodguard"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/goprintffuncname"
@@ -370,6 +370,10 @@ func (LinterBuilder) Build(cfg *config.Config) ([]*linter.Config, error) {
 			WithAutoFix().
 			WithURL("https://github.com/segmentio/golines"),
 
+		linter.NewConfig(golistics.New()).
+			WithSince("v2.9.0").
+			WithURL("https://github.com/ufukty/golistics"),
+
 		linter.NewConfig(goheader.New(&cfg.Linters.Settings.Goheader, placeholderReplacer)).
 			WithSince("v1.28.0").
 			WithAutoFix().
@@ -388,10 +392,6 @@ func (LinterBuilder) Build(cfg *config.Config) ([]*linter.Config, error) {
 			WithSince("v2.6.0").
 			WithLoadForGoAnalysis().
 			WithURL("https://pkg.go.dev/golang.org/x/tools/go/analysis/passes/modernize"),
-
-		linter.NewConfig(gomethods.New()).
-			WithSince("v2.9.0").
-			WithURL("https://github.com/ufukty/gomethods"),
 
 		linter.NewConfig(gomoddirectives.New(&cfg.Linters.Settings.GoModDirectives)).
 			WithSince("v1.39.0").
